+++ fv3core/pace/fv3core/stencils/map_single.py
@@ -5,7 +5,7 @@ from gt4py.cartesian.gtscript import FORWARD, PARALLEL, computation, interval
 import pace.util
 from pace.dsl.dace import orchestrate
 from pace.dsl.stencil import StencilFactory
-from pace.dsl.typing import FloatField, FloatFieldIJ, IntFieldIJ  # noqa: F401
+from pace.dsl.typing import Float, FloatField, FloatFieldIJ, IntFieldIJ  # noqa: F401
 from pace.fv3core.stencils.basic_operations import copy_defn
 from pace.fv3core.stencils.remap_profile import RemapProfile
 from pace.util import X_DIM, Y_DIM, Z_DIM
@@ -103,14 +103,22 @@ class MapSingle:
         grid_indexing = stencil_factory.grid_indexing
 
         def make_quantity():
-            return quantity_factory.zeros([X_DIM, Y_DIM, Z_DIM], units="unknown")
+            return quantity_factory.zeros(
+                [X_DIM, Y_DIM, Z_DIM],
+                units="unknown",
+                dtype=Float,
+            )
 
         self._dp1 = make_quantity()
         self._q4_1 = make_quantity()
         self._q4_2 = make_quantity()
         self._q4_3 = make_quantity()
         self._q4_4 = make_quantity()
-        self._tmp_qs = quantity_factory.zeros([X_DIM, Y_DIM], units="unknown")
+        self._tmp_qs = quantity_factory.zeros(
+            [X_DIM, Y_DIM],
+            units="unknown",
+            dtype=Float,
+        )
         self._lev = quantity_factory.zeros([X_DIM, Y_DIM], units="", dtype=int)
 
         self._copy_stencil = stencil_factory.from_dims_halo(
@@ -150,7 +158,7 @@ class MapSingle:
         pe1: FloatField,
         pe2: FloatField,
         qs: Optional["FloatFieldIJ"] = None,
-        qmin: float = 0.0,
+        qmin: Float = 0.0,
     ):
         """
         Compute x-flux using the PPM method.
