+++ fv3core/pace/fv3core/stencils/neg_adj3.py
@@ -4,7 +4,7 @@ from gt4py.cartesian.gtscript import BACKWARD, FORWARD, PARALLEL, computation, i
 import pace.util
 import pace.util.constants as constants
 from pace.dsl.stencil import StencilFactory
-from pace.dsl.typing import FloatField, FloatFieldIJ
+from pace.dsl.typing import Float, FloatField, FloatFieldIJ
 from pace.util import X_DIM, Y_DIM
 
 
@@ -100,8 +100,8 @@ def fix_neg_water(
     qsnow: FloatField,
     qice: FloatField,
     qgraupel: FloatField,
-    lv00: float,
-    d0_vap: float,
+    lv00: Float,
+    d0_vap: Float,
 ):
     """
     Args:
@@ -340,8 +340,16 @@ class AdjustNegativeTracerMixingRatio:
         hydrostatic: bool,
     ):
         grid_indexing = stencil_factory.grid_indexing
-        self._sum1 = quantity_factory.zeros([X_DIM, Y_DIM], units="unknown")
-        self._sum2 = quantity_factory.zeros([X_DIM, Y_DIM], units="unknown")
+        self._sum1 = quantity_factory.zeros(
+            [X_DIM, Y_DIM],
+            units="unknown",
+            dtype=Float,
+        )
+        self._sum2 = quantity_factory.zeros(
+            [X_DIM, Y_DIM],
+            units="unknown",
+            dtype=Float,
+        )
         if check_negative:
             raise NotImplementedError(
                 "Unimplemented namelist value check_negative=True"
